qa/src/sizeof.c: use %zd and sizeof(...) instead of %d and (int)sizeof(...)

+++ qa/src/sizeof.c
@@ -13,21 +13,21 @@ main(int argc, char **argv)
 {
     while (argc > 1) {
 	if (strcmp(argv[1], "char") == 0)
-	    printf("char=%d\n", (int)sizeof(char));
+	    printf("char=%zd\n", sizeof(char));
 	else if (strcmp(argv[1], "short") == 0)
-	    printf("short=%d\n", (int)sizeof(short));
+	    printf("short=%zd\n", sizeof(short));
 	else if (strcmp(argv[1], "int") == 0)
-	    printf("int=%d\n", (int)sizeof(int));
+	    printf("int=%zd\n", sizeof(int));
 	else if (strcmp(argv[1], "long") == 0)
-	    printf("long=%d\n", (int)sizeof(long));
+	    printf("long=%zd\n", sizeof(long));
 	else if (strcmp(argv[1], "long long") == 0)
-	    printf("long long=%d\n", (int)sizeof(long long));
+	    printf("long long=%zd\n", sizeof(long long));
 	else if (strcmp(argv[1], "float") == 0)
-	    printf("float=%d\n", (int)sizeof(float));
+	    printf("float=%zd\n", sizeof(float));
 	else if (strcmp(argv[1], "double") == 0)
-	    printf("double=%d\n", (int)sizeof(long));
+	    printf("double=%zd\n", sizeof(long));
 	else if (strcmp(argv[1], "ptr") == 0)
-	    printf("ptr=%d\n", (int)sizeof(void *));
+	    printf("ptr=%zd\n", sizeof(void *));
 	else {
 	    printf("Botch: don't grok type \"%s\"\n", argv[1]);
 	    return 1;
